requirements.txt
torch>=1.8.0
torchvision>=0.9.0
timm>=0.5.4
numpy>=1.19.5
pandas>=1.2.4
matplotlib>=3.4.2
seaborn>=0.11.1
scikit-learn>=0.24.2
opencv-python>=4.5.2
tqdm>=4.61.0
Pillow>=8.2.0
thop>=0.0.31-2005241907
jupyter>=1.0.0
tensorboard>=2.5.0